fix notes

+++ src/services/fleet_manager.py
@@ -243,32 +243,38 @@ class FleetManager:
     def active_user_ids(self) -> list[int]:
         return sorted(self.active_rides.active_user_ids())
 
-    def apply_treatment(self, location:tuple[float, float]) -> list[str]:
+    def apply_treatment(self, treatment_location:tuple[float, float]) -> list[str]:
         """
         Run maintenance on all treatable vehicles
         Arg:
-            location (tuple[float, float]): The (latitude, longitude) of the user.
+            location (tuple[float, float]): The treatment location (latitude, longitude).
         Return:
                 vehicle IDs list of treated vehicles
         """
         treated: list[str] = []
-        # for eligible vehicles
-        for vehicle_id, vehicle in self.vehicles.items():
-                if vehicle.can_initiate_treatment():
-                    vehicle.apply_treatment()
-                    treated.append(vehicle_id)
+        degraded_ids = set(self.degraded_repo.get_vehicle_ids())
+        non_degraded_ids = [vid for vid in self.vehicles.keys()
+                            if vid not in degraded_ids
+                            and self.vehicles[vid].can_initiate_treatment()]
 
         # for degraded vehicles
-        degraded_vehicles = self.degraded_repo.get_vehicle_ids()
-        for vehicle_id in degraded_vehicles:
-            nearest_station = self._nearest_station_with_free_slot(location=location)
-            vehicle = self.vehicles[vehicle_id]
+        for vehicle_id in degraded_ids:
+            nearest_station = self._nearest_station_with_free_slot(location=treatment_location)
             if nearest_station is None:
                 raise ConflictError("No station with free slot available")
-            vehicle.apply_treatment()
+            degr_vehicle = self.vehicles.get(vehicle_id)
+            if degr_vehicle is None:
+                raise NotFoundError(f"Vehicle {vehicle_id} not found.")
+            degr_vehicle.apply_treatment()
             self.degraded_repo.remove_vehicle(vehicle_id)
             nearest_station.add_vehicle(vehicle_id)
-            vehicle.dock_to_station(nearest_station.container_id)
+            degr_vehicle.dock_to_station(nearest_station.container_id)
+            treated.append(vehicle_id)
+
+        # for eligible vehicles that can have a treatment
+        for vehicle_id in non_degraded_ids:
+            vehicle = self.vehicles[vehicle_id]
+            vehicle.apply_treatment()
             treated.append(vehicle_id)
 
         return treated
